+++ src/pyproject.toml
@@ -19,6 +19,7 @@ html5lib = "^1.1"
 pyexploitdb = "^0.2.9"
 pyserde = "^0.12.3"
 orjson = "^3.9.9"
+pandas = "^2.1.3"
 
 
 [build-system]
